Refactor: Move unit test serialization to new class
Extract unit test serialization, result parsing, and formatting logic from UnitTestSystem into a new static UnitTestSerializer class. Update all usages in UnitTestSystem, UnitTestEngine, and UnitTestBisectorSystem to use UnitTestSerializer, improving code organization and separation of concerns.
